+++ lib/dartapi_core.dart
@@ -11,6 +11,8 @@ export 'src/core/background_task.dart';
 export 'src/core/websocket_route.dart';
 export 'src/core/pagination.dart';
 export 'src/core/sse.dart';
+export 'src/core/env_config.dart';
+export 'src/core/health_controller.dart';
 
 export 'src/openapi/security_scheme.dart';
 export 'src/openapi/openapi_generator.dart';